Contact legal authority changes

+++ visitz/VisitzModel/Models/People/ContactLegalAuthority.cs
@@ -190,13 +190,15 @@ public partial class ContactLegalAuthority : IRealmObject, IApiJson<ContactLegal
         );
     }
 
-    public static void RemoveByParent(Realm realm, List<string> parentContactIdList)
+    public static void RemoveByParent(Realm realm, string parentContactId)
     {
-        foreach (var contactId in parentContactIdList)
+        var contacts = realm.All<IcmContact>().Where(item => item.Id == parentContactId);
+
+        if (contacts.Count() <= 1)
         {
             var contactLegalAuthorityToBeDeleted = realm
                 .All<ContactLegalAuthority>()
-                .Where(item => item.ParentContactId == contactId);
+                .Where(item => item.ParentContactId == parentContactId);
 
             realm.RemoveRange(contactLegalAuthorityToBeDeleted);
         }
